Move `get_branch_status` and its helpers to `StatusMixin`
